divide functions in 2 files
